+++ example/react-native/android/app/build.gradle
@@ -137,7 +137,7 @@ android {
 }
 
 dependencies {
-    compile project(':aerogear-react-native-core')
+    compile project(':aerogear-react-native-sdk')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
