+++ example/react-native/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'AeroGearDemo'
-include ':aerogear-react-native-core'
-project(':aerogear-react-native-core').projectDir = new File(rootProject.projectDir, '../node_modules/aerogear-react-native-core/android')
+include ':aerogear-react-native-sdk'
+project(':aerogear-react-native-sdk').projectDir = new File(rootProject.projectDir, '../node_modules/aerogear-react-native-sdk/android')
 
 include ':app'
