+++ example/react-native/package.json
@@ -7,7 +7,7 @@
     "test": "jest"
   },
   "dependencies": {
-    "aerogear-react-native-core": "file:../../../aerogear-react-native-core/",
+    "aerogear-react-native-sdk": "file:../../../aerogear-react-native-sdk/",
     "react": "^16.2.0",
     "react-native": "0.53.3",
     "react-native-navbar": "^2.1.0",
