+++ example/react-native/src/components/screens/HomeScreen.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { View, Text, StatusBar } from "react-native";
-import { RNMetricsService } from "aerogear-react-native-core";
+import { RNMetricsService } from "aerogear-react-native-sdk";
 import { NavDrawerButton } from "../common";
 import { Colors } from "../../assets";
