+++ packages/core/package.json
@@ -1,10 +1,15 @@
 {
-  "name": "@aerogearservices/core",
+  "name": "aerogear-core",
   "version": "1.0.0",
-  "description": "Core SDK for AeroGear services",
+  "description": "JavaScript Core SDK for AeroGear services",
   "main": "dist/index.js",
   "types": "./types/main.d.ts",
   "author": "Aerogear <aerogear@googlegroups.com>",
+  "keywords": [
+    "react-native",
+    "aerogear",
+    "mobile"
+  ],
   "license": "Apache-2.0",
   "scripts": {
     "clean": "del coverage src/**/*.js src/**/*.map test/**/*.js test/**/*.map",
